refactor: split pedido request models
