+++ src/LexosHub.ERP.VarejOnline.Infra.ErpApi/Request/Pedido/RepresentanteRef.cs
@@ -1,0 +1,13 @@
+using Newtonsoft.Json;
+
+namespace LexosHub.ERP.VarejOnline.Infra.ErpApi.Request.Pedido
+{
+    public class RepresentanteRef
+    {
+        [JsonProperty("id")]
+        public long? Id { get; set; }
+
+        [JsonProperty("nome")]
+        public string? Nome { get; set; }
+    }
+}
