+++ src/LexosHub.ERP.VarejOnline.Infra.ErpApi/Request/Pedido/VoucherPagamento.cs
@@ -1,0 +1,13 @@
+using Newtonsoft.Json;
+
+namespace LexosHub.ERP.VarejOnline.Infra.ErpApi.Request.Pedido
+{
+    public class VoucherPagamento
+    {
+        [JsonProperty("voucher")]
+        public VoucherRef? Voucher { get; set; }
+
+        [JsonProperty("valor")]
+        public decimal? Valor { get; set; }
+    }
+}
